requirements.txt
# API Framework
fastapi>=0.68.0
uvicorn>=0.15.0

# Database
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.1
alembic>=1.7.0

# AI/ML
langchain>=0.0.200
openai>=0.27.0

# Data Validation
pydantic>=1.8.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5

# Testing
pytest>=6.2.5
pytest-asyncio>=0.15.1
httpx>=0.18.2

# Code Quality
black>=21.7b0
flake8>=3.9.0
pre-commit>=2.15.0

# Utilities
python-dotenv>=0.19.0
requests>=2.26.0
aiohttp>=3.8.0

# Data Processing
pandas>=1.3.0
numpy>=1.21.0

# Visualization (for admin dashboard)
plotly>=5.1.0

# Development Tools
jupyter>=1.0.0
ipython>=7.27.0 